+++ src/api/axiosApi/bizApi.js
@@ -1,0 +1,3 @@
+export const Biz = {
+  BizSearchList: "/api/manage-user/bizListBody.do",
+};
