+++ src/components/hook/biz/useBizListSearchQuery.js
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/vue-query";
+import { bizListSearchApi } from "../../../api/biz/bizListSearchApi";
+
+export const useBizListSearchQuery = (injectedValue, cPage) => {
+  return useQuery({
+    queryKey: ["bizList", injectedValue, cPage],
+    queryFn: () => bizListSearchApi(injectedValue, cPage),
+  });
+};
