+++ src/components/page/Manage-user/BizMain.vue
@@ -39,38 +39,21 @@
         </template>
       </tbody>
     </table>
-    <Pagination
-      :totalItems="bizList?.bizCnt || 0"
-      :items-per-page="5"
-      :max-pages-shown="5"
-      :onClick="searchList"
-      v-model="cPage"
-    />
+    <Pagination :totalItems="bizList?.bizCnt || 0" :items-per-page="5" :max-pages-shown="5" :onClick="searchList"
+      v-model="cPage" />
   </div>
 </template>
 
 <script setup>
-import { useQuery } from "@tanstack/vue-query";
-import axios from "axios";
 import { useModalStore } from "../../../stores/modalState";
 import Pagination from "../../common/Pagination.vue";
+import { useBizListSearchQuery } from "../../hook/biz/useBizListSearchQuery";
+
 
 const cPage = ref(1);
 const modalStateBiz = useModalStore();
 const bizIdx = ref("");
 const injectedValue = inject("provideValue");
-
-const searchList = async () => {
-  const data = {
-    ...injectedValue.value,
-    currentPage: cPage.value.toString(),
-    pageSize: (5).toString(),
-  };
-
-  const result = await axios.post("/api/manage-user/bizListBody.do", data);
-
-  return result.data;
-};
 
 const {
   data: bizList,
@@ -78,10 +61,7 @@ const {
   isSuccess,
   isError,
   refetch,
-} = useQuery({
-  queryKey: ["bizList", injectedValue, cPage],
-  queryFn: searchList,
-});
+} = useBizListSearchQuery(injectedValue, cPage);
 
 const handlerModal = (idx) => {
   bizIdx.value = idx;
